// src/services/feedly/types.ts
export interface ExchangeTokenResponse {
    id: string;
    access_token: string;
    refresh_token: string;
    expires_in: number;
    token_type: string;
    plan: string;
    state?: string;
}
// ...
export interface Credential extends ExchangeTokenResponse {
    authorized: number;
}

export interface Environment {
    client_id: string;
    client_secret: string;
    scope: string;
    redirect_uri: string;
}